old API deprecated and moved to eCAL::v5
